Release 5.0.0 (#9736)
Replace Fusions with Structural Variants (not backwards compatible)

* Remove support for FUSION datatype.
* bug fix in TabDelimitedFileUtil.java parseInt was using Float.parseFloat and the casting to int and losing precision (string "138536968" was converted to int 138536960)
* update mybatis xml (#24)
* fix sv gene counts (now using unique identifier instead of stable id)
* prevent double counting of sv records where SITE1 and SITE2 genes are the same
* add germline to structural variant table and model
* changed IS_GERMLINE to SV_STATUS
* accept NULL values for SV data values
* adjust mybatis mapper to work with NULL
* take out ensembl check
* first pass at modifying migration script to not allow fusions
* change import gene panel matrix to use WGS WXS and NA placeholders
* update migration step version in migrate_db
* update version numbers for db
* comment the core module StructuralVariant model with warnings
* add support for WXS/WGS
* use StableIdUtil getSample to convert sampleid before looking up in db
* only add a unique set of records
* update length field name and test study files
* update test files
* updating unique sv record
* add SV changes to validator
* Validate Sample IDs in SV class
* fix the logic for finding cases without panel data
* Add documentation for session service management

+++ core/src/main/java/org/mskcc/cbio/maf/TabDelimitedFileUtil.java
@@ -1,15 +1,15 @@
 /*
- * Copyright (c) 2015 Memorial Sloan-Kettering Cancer Center.
+ * Copyright (c) 2015 - 2022 Memorial Sloan Kettering Cancer Center.
  *
  * This library is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY, WITHOUT EVEN THE IMPLIED WARRANTY OF MERCHANTABILITY OR FITNESS
  * FOR A PARTICULAR PURPOSE. The software and documentation provided hereunder
- * is on an "as is" basis, and Memorial Sloan-Kettering Cancer Center has no
+ * is on an "as is" basis, and Memorial Sloan Kettering Cancer Center has no
  * obligations to provide maintenance, support, updates, enhancements or
- * modifications. In no event shall Memorial Sloan-Kettering Cancer Center be
+ * modifications. In no event shall Memorial Sloan Kettering Cancer Center be
  * liable to any party for direct, indirect, special, incidental or
  * consequential damages, including lost profits, arising out of the use of this
- * software and its documentation, even if Memorial Sloan-Kettering Cancer
+ * software and its documentation, even if Memorial Sloan Kettering Cancer
  * Center has been advised of the possibility of such damage.
  */
 
@@ -141,11 +141,19 @@ public class TabDelimitedFileUtil
 		}
 	}
 
+	// This method does not call Integer.parseInt() as one might expect.
+	// Presumably this is to allow the convertion of strings like "6.2" to "6".
+	// The method previously called (int)Float.parseFloat() but floats
+	// reserve 23 bits for the mantissa and ints are 32 bits so precision
+	// was lost parsing "138536968" which was converted to 138536960. 
+	// Now we call (int)Double.parseDouble() because double allocates
+	// 52 bits for the mantissa. Note getPartLong calls Long.parseLong()
+	// when this does not call Integer.parseInt() which seems inconsistent.
 	public static Integer getPartInt(int index, String[] parts)
 	{
 		try {
 			String part = parts[index];
-			return (int)(Float.parseFloat(part));
+			return (int)(Double.parseDouble(part));
 		} catch (ArrayIndexOutOfBoundsException e) {
 			return NA_INT;
 		} catch (NumberFormatException e) {
